+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/mutualBlock/MutualBlockItem.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.main;
+package nl.tudelft.cs4160.trustchain_android.mutualBlock;
 
 /**
  * Class that is used to define a MutualBlock in the recycler view.
